Move writing to a separate module in drcp-format
