+++ drcp-format/src/lib.rs
@@ -2,19 +2,12 @@
 //! (DRCP) proof format. The format can be used by Constraint Programming solvers to provide a
 //! certifyable proof of unsatisfiability or optimality.
 //!
-//! To write DRCP proof files, look at the [`ProofWriter`] documentation. Literal definition files
-//! can be written using [`LiteralDefinitions`].
+//! To write DRCP proof files, look at the [`writer::ProofWriter`] documentation. Literal definition
+//! files can be written using [`writer::LiteralDefinitions`].
 
-mod atomic;
-mod encountered_literals;
 mod format;
-mod literal_code_provider;
-mod writer;
+pub mod writer;
 
 pub mod steps;
 
-pub use atomic::*;
-pub use encountered_literals::*;
 pub use format::*;
-pub use literal_code_provider::*;
-pub use writer::*;
