+++ drcp-format/src/writer/atomic.rs
