+++ drcp-format/src/writer/literal_code_provider.rs
